+++ pkg/model/InterestMapping.go
@@ -2,5 +2,5 @@ package model
 
 type InterestMapping struct {
 	UserID     uint64
-	InterestID uint64
+	InterestID uint
 }
